+++ src/config/app.php
@@ -192,7 +192,7 @@ return [
 
     'aliases' => Facade::defaultAliases()->merge(
         [
-        // ...
+            // ...
         ]
     )->toArray(),