+++ src/routes/web.php
@@ -14,7 +14,7 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', 'DashboardController@dashboard')->name('harvey');
-Route::get('/harvey-pipeline', 'DashboardController@readPipeline')->name('harvey-pipeline');
+Route::get('/harvey-deployment', 'DashboardController@readDeployment')->name('harvey-deployment');
 Route::get('/harvey-project', 'DashboardController@readProject')->name('harvey-project');
 
 Route::post('/unlock-project', 'DashboardController@unlockProject');