+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 
-## NEXT RELEASE
+## v0.2.0 (2022-05-01)
 
+- Renames all `pipeline` references to `deployment` to match the Harvey backend
 - Adds authentication to Harvey API with basic auth when present
 - Adds `lock` and `unlock` functionality for projects
 - Wraps all HTTP requests in try/catch blocks to gracefully handle exceptions
@@ -13,5 +14,5 @@
 
 - Initial release
 - Can view a list of projects from Harvey
-- Can view a list of pipelines per project and their status and output
+- Can view a list of deployments per project and their status and output
 - Scaffolding for future features